refactor(model.Car) : 기존 Application class 내에 정의되어있던 Car class 를 model로 분리
기존에 Application 내에 정의되어있던 Car class 를 model로 분리시켜 MVC 패턴을 지키도록 코드를 리팩토링 했습니다.
자동차 이름이 5글자를 초과했을때 IllegalArgumentException 을 던지도록 하는 로직은 Car 의 constructor에서 처리할 수 있도록 변경했습니다.
